requirements.txt
aioredis==1.2.0
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
awesome-slugify==1.6.5
certifi==2019.6.16
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
demjson==2.2.4
Django==2.1.7
django-annoying==0.10.6
django-braces==1.13.0
django-filter==2.2.0
django-haystack==2.8.1
django-jsonstore==0.4.1
django-material==1.6.0
django-model-utils==3.2.0
django-mysql==3.3.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-viewflow==1.6.0
djangorestframework==3.10.2
flatlib==0.2.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jieba==0.39
jsonfield==2.0.2
Markdown==3.1
msgpack==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.2
pdfminer==20191103
Pillow==6.0.0
pycparser==2.19
pycryptodome==3.9.1
PyHamcrest==1.9.0
PyJWT==1.7.1
PyPDF2==1.26.0
pyswisseph==2.0.0.post2
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
redis==3.2.1
regex==2019.3.12
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
Unidecode==0.4.21
urllib3==1.25.3
uWSGI==2.0.18
virtualenv==16.7.6
Whoosh==2.7.4
zope.interface==4.6.0
